feat: Complete Phase 1 - Investment Creation
- Add investment entity with DB schema (007_investments.sql)
- Implement POST /investments endpoint with ledger posting
- Emit investment.created events via RabbitMQ
- Refactor HTTP handlers to shared cmd/httpHandler
- Add investor capital account seed migration
- Fix annual rate handling (hardcoded at 0.30)
- Update transport layer with publisher wiring

+++ internal/domain/ledger_test.go
@@ -89,9 +89,9 @@ func TestPostTransaction(t *testing.T) {
 	if err != nil {
 		t.Fatalf("Failed to get balance: %v", err)
 	}
-	expectedBalance := int64(100)
+	expectedBalance := 100.0
 	if balance != expectedBalance {
-		t.Errorf("Expected balance %d, got %d", expectedBalance, balance)
+		t.Errorf("Expected balance %f, got %f", expectedBalance, balance)
 	}
 
 	// Test get account history
